+++ lib/text.ts
@@ -3,8 +3,6 @@ import {
 	removeComment,
 	escape,
 
-	/* PRINT ONLY END */
-
 	/* NOT FOR BROWSER */
 
 	sanitize,