+++ src/attributes.ts
@@ -14,7 +14,16 @@ import {Token} from './index';
 import {AtomToken} from './atom';
 import {AttributeToken} from './attribute';
 import type {Config, LintError} from '../base';
-import type {ExtToken, HtmlToken, SyntaxToken, TdToken} from '../internal';
+import type {
+	ExtToken,
+	HtmlToken,
+	SyntaxToken,
+	TdToken,
+
+	/* NOT FOR BROWSER */
+
+	ParameterToken,
+} from '../internal';
 import type {AttributeTypes} from './attribute';
 import type {TableTokens} from './table/index';
 
@@ -428,7 +437,7 @@ export abstract class AttributesToken extends Token {
 		} else {
 			token.close();
 		}
-		if (this.closest('parameter')) {
+		if (this.closest('parameter,ext')?.is<ParameterToken>('parameter')) {
 			token.escape();
 		}
 		super.insertAt(token, i);