+++ src/table/index.ts
@@ -16,7 +16,16 @@ import type {
 	LintError,
 	AST,
 } from '../../base';
-import type {AttributesToken, TdToken, TrToken, Token} from '../../internal';
+import type {
+	AttributesToken,
+	TdToken,
+	TrToken,
+	Token,
+
+	/* NOT FOR BROWSER */
+
+	ParameterToken,
+} from '../../internal';
 import type {TableCoords} from './trBase';
 
 /* NOT FOR BROWSER */
@@ -122,7 +131,9 @@ export abstract class TableToken extends TrBaseToken {
 
 	set closed(closed) {
 		if (closed && !this.closed) {
-			this.close(this.closest('parameter-value') ? '\n{{!}}}' : '\n|}');
+			this.close(
+				this.closest('parameter,ext')?.is<ParameterToken>('parameter') ? '\n{{!}}}' : '\n|}',
+			);
 		}
 	}